feature : add permission handler class
